refactor: 拆分设置页专家管理
